# openapi.yaml
openapi: 3.1.0
info:
  title: Managed Agent V1 Platform
  version: 1.0.0
servers:
  - url: http://localhost:8000
security:
  - InstanceToken: []
paths:
  /v1/instances/register:
    post:
      tags: [instances]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInstanceRequest"
      responses:
        "201":
          description: Instance registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterInstanceResponse"
  /v1/orchestrations:
    post:
      tags: [orchestrations]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Orchestration"
      responses:
        "201":
          description: Orchestration published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orchestration"
  /v1/orchestrations/sync:
    get:
      tags: [orchestrations]
      responses:
        "200":
          description: Synced orchestrations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncOrchestrationsResponse"
  /v1/orchestrations/{orchestration_id}:
    get:
      tags: [orchestrations]
      parameters:
        - in: path
          name: orchestration_id
          required: true
          schema:
            type: string
      security: []
      responses:
        "200":
          description: Orchestration detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orchestration"
  /v1/runs:
    post:
      tags: [runs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRunRequest"
      responses:
        "201":
          description: Run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartRunResponse"
  /v1/runs/{run_id}:
    get:
      tags: [runs]
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardRunDetail"
  /v1/tasks/{task_id}:
    get:
      tags: [runs]
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunTask"
  /v1/runs/{run_id}/events:batch:
    post:
      tags: [runs]
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRunEventsRequest"
      responses:
        "202":
          description: Events accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventBatchAcceptedResponse"
  /v1/policy/preauthorize:
    post:
      tags: [policy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreauthorizeActionRequest"
      responses:
        "200":
          description: Policy decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDecision"
  /v1/policy-decisions/{decision_id}:
    get:
      tags: [policy]
      parameters:
        - in: path
          name: decision_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy decision detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDecision"
  /v1/policy-decisions/{decision_id}/resolve:
    post:
      tags: [policy]
      security: []
      parameters:
        - in: path
          name: decision_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyDecisionResolutionRequest"
      responses:
        "200":
          description: Policy decision resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDecision"
  /v1/runs/{run_id}/complete:
    post:
      tags: [runs]
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteRunRequest"
      responses:
        "200":
          description: Run completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationStatusResponse"
  /v1/dashboard/overview:
    get:
      tags: [dashboard]
      security: []
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverviewResponse"
  /v1/dashboard/orchestrations:
    get:
      tags: [dashboard]
      security: []
      responses:
        "200":
          description: Dashboard orchestrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Orchestration"
  /v1/dashboard/instances:
    get:
      tags: [dashboard]
      security: []
      responses:
        "200":
          description: Dashboard instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodexInstance"
  /v1/dashboard/runs:
    get:
      tags: [dashboard]
      security: []
      responses:
        "200":
          description: Dashboard runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
  /v1/dashboard/runs/{run_id}:
    get:
      tags: [dashboard]
      security: []
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard HTML detail
          content:
            text/html:
              schema:
                type: string
components:
  securitySchemes:
    InstanceToken:
      type: apiKey
      in: header
      name: X-Instance-Token
  schemas:
    RegisterInstanceRequest:
      type: object
      required: [client_kind, workspace_path, capabilities, machine_id]
      properties:
        client_kind: { type: string }
        workspace_path: { type: string }
        capabilities:
          type: array
          items: { type: string }
        machine_id: { type: string }
    RegisterInstanceResponse:
      type: object
      required: [instance_id, instance_token, registered_at]
      properties:
        instance_id: { type: string }
        instance_token: { type: string }
        registered_at: { type: string, format: date-time }
    CodexInstance:
      type: object
      required: [instance_id, machine_id, client_kind, workspace_path, capabilities, registered_at, last_seen_at, instance_token]
      properties:
        instance_id: { type: string }
        machine_id: { type: string }
        client_kind: { type: string }
        workspace_path: { type: string }
        capabilities:
          type: array
          items: { type: string }
        registered_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        instance_token: { type: string }
    Orchestration:
      type: object
      required: [orchestration_id, name, version, status, entrypoint, required_tools, required_skills, policy_profile, compatibility]
      properties:
        orchestration_id: { type: string }
        name: { type: string }
        version: { type: string }
        status: { type: string }
        entrypoint: { type: string }
        required_tools:
          type: array
          items: { type: string }
        required_skills:
          type: array
          items: { type: string }
        policy_profile: { type: string }
        compatibility:
          type: array
          items: { type: string }
        published_at:
          type: string
          format: date-time
          nullable: true
    SyncOrchestrationsResponse:
      type: object
      required: [orchestrations, synced_at]
      properties:
        orchestrations:
          type: array
          items:
            $ref: "#/components/schemas/Orchestration"
        synced_at: { type: string, format: date-time }
    StartRunRequest:
      type: object
      required: [orchestration_id, instance_id, title, goal, workspace_path, trigger]
      properties:
        orchestration_id: { type: string }
        instance_id: { type: string }
        title: { type: string }
        goal: { type: string }
        workspace_path: { type: string }
        trigger: { type: string }
    StartRunResponse:
      type: object
      required: [run_id, task_id, policy_profile, started_at]
      properties:
        run_id: { type: string }
        task_id: { type: string }
        policy_profile: { type: string }
        started_at: { type: string, format: date-time }
    Run:
      type: object
      required: [run_id, orchestration_id, orchestration_version, instance_id, status, started_at, trigger, workspace_path]
      properties:
        run_id: { type: string }
        orchestration_id: { type: string }
        orchestration_version: { type: string }
        instance_id: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
        trigger: { type: string }
        workspace_path: { type: string }
        summary: { type: string, nullable: true }
    RunTask:
      type: object
      required: [task_id, run_id, title, goal, status, started_at]
      properties:
        task_id: { type: string }
        run_id: { type: string }
        title: { type: string }
        goal: { type: string }
        status: { type: string }
        current_step: { type: string, nullable: true }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
    RunEvent:
      type: object
      required: [event_id, run_id, task_id, source, type, timestamp, payload]
      properties:
        event_id: { type: string }
        run_id: { type: string }
        task_id: { type: string }
        source: { type: string }
        type: { type: string }
        timestamp: { type: string, format: date-time }
        payload:
          type: object
          additionalProperties: true
    BatchRunEventsRequest:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/RunEvent"
    EventBatchAcceptedResponse:
      type: object
      required: [accepted]
      properties:
        accepted: { type: integer }
    PreauthorizeActionRequest:
      type: object
      required: [run_id, task_id, action_type, target, workspace_path, tool_name, metadata]
      properties:
        run_id: { type: string }
        task_id: { type: string }
        action_type: { type: string }
        target: { type: string }
        workspace_path: { type: string }
        tool_name: { type: string }
        metadata:
          type: object
          additionalProperties: true
    PolicyDecision:
      type: object
      required: [decision_id, run_id, task_id, action_type, decision, reason, status, requested_at]
      properties:
        decision_id: { type: string }
        run_id: { type: string }
        task_id: { type: string }
        action_type: { type: string }
        decision: { type: string }
        reason: { type: string }
        status: { type: string }
        requested_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        resolved_by: { type: string, nullable: true }
        comment: { type: string, nullable: true }
    PolicyDecisionResolutionRequest:
      type: object
      required: [resolution, resolved_by]
      properties:
        resolution: { type: string }
        resolved_by: { type: string }
        comment: { type: string, nullable: true }
    CompleteRunRequest:
      type: object
      required: [status, ended_at]
      properties:
        status: { type: string }
        summary: { type: string, nullable: true }
        ended_at: { type: string, format: date-time }
    OperationStatusResponse:
      type: object
      required: [status]
      properties:
        status: { type: string }
    ToolExecution:
      type: object
      required: [tool_execution_id, run_id, task_id, tool_name, status, started_at]
      properties:
        tool_execution_id: { type: string }
        run_id: { type: string }
        task_id: { type: string }
        tool_name: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
        input_summary: { type: string, nullable: true }
        output_summary: { type: string, nullable: true }
        error_summary: { type: string, nullable: true }
    WorkflowFingerprint:
      type: object
      required: [fingerprint_id, title_pattern, orchestration_id, step_signature, tool_signature, occurrence_count, last_seen_at, terminal_status, error_categories]
      properties:
        fingerprint_id: { type: string }
        title_pattern: { type: string }
        orchestration_id: { type: string }
        step_signature:
          type: array
          items: { type: string }
        tool_signature:
          type: array
          items: { type: string }
        occurrence_count: { type: integer }
        last_seen_at: { type: string, format: date-time }
        terminal_status: { type: string }
        error_categories:
          type: array
          items: { type: string }
    DashboardRunDetail:
      type: object
      required: [run, task, events, tool_executions, policy_decisions]
      properties:
        run:
          $ref: "#/components/schemas/Run"
        task:
          $ref: "#/components/schemas/RunTask"
        events:
          type: array
          items:
            $ref: "#/components/schemas/RunEvent"
        tool_executions:
          type: array
          items:
            $ref: "#/components/schemas/ToolExecution"
        policy_decisions:
          type: array
          items:
            $ref: "#/components/schemas/PolicyDecision"
        fingerprint:
          $ref: "#/components/schemas/WorkflowFingerprint"
        orchestration:
          $ref: "#/components/schemas/Orchestration"
        instance:
          $ref: "#/components/schemas/CodexInstance"
    DashboardOverviewResponse:
      type: object
      required: [orchestration_count, active_instance_count, running_run_count, blocked_run_count, recent_error_count, pending_policy_decision_count, top_workflows, error_breakdown]
      properties:
        orchestration_count: { type: integer }
        active_instance_count: { type: integer }
        running_run_count: { type: integer }
        blocked_run_count: { type: integer }
        recent_error_count: { type: integer }
        pending_policy_decision_count: { type: integer }
        top_workflows:
          type: array
          items:
            $ref: "#/components/schemas/DashboardWorkflowSummary"
        error_breakdown:
          type: array
          items:
            $ref: "#/components/schemas/DashboardErrorSummary"
    DashboardWorkflowSummary:
      type: object
      required: [fingerprint_id, orchestration_id, title_pattern, occurrence_count, terminal_status, error_categories, last_seen_at]
      properties:
        fingerprint_id: { type: string }
        orchestration_id: { type: string }
        title_pattern: { type: string }
        occurrence_count: { type: integer }
        terminal_status: { type: string, enum: [completed, failed] }
        error_categories:
          type: array
          items: { type: string }
        last_seen_at: { type: string, format: date-time }
    DashboardErrorSummary:
      type: object
      required: [category, occurrence_count, affected_orchestration_ids, last_seen_at]
      properties:
        category: { type: string }
        occurrence_count: { type: integer }
        affected_orchestration_ids:
          type: array
          items: { type: string }
        last_seen_at: { type: string, format: date-time }